refactor(ohunt): data -> database
