+++ ohunt/OHunt.Web/Controllers/SubmissionsController.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.AspNet.OData;
 using Microsoft.AspNetCore.Mvc;
-using OHunt.Web.Data;
+using OHunt.Web.Database;
 using OHunt.Web.Errors;
 using OHunt.Web.Models;
